moved MicroClimateZone class to own file
