+++ Models/MicroClimate/MicroClimateZone.cs
@@ -1,0 +1,191 @@
+﻿using System;
+using APSIM.Shared.Utilities;
+using System.Collections.Generic;
+using Models.Core;
+
+namespace Models
+{
+    public partial class MicroClimate
+    {
+        /// <summary>
+        /// MicroClimateZone
+        /// </summary>
+        [Serializable]
+        private class MicroClimateZone
+        {
+            public Zone zone = null;
+
+            /// <summary>The _albedo</summary>
+            public double _albedo = 0;
+
+            /// <summary>The net long wave</summary>
+            public double netLongWave;
+
+            /// <summary>The sum rs</summary>
+            public double sumRs;
+
+            /// <summary>The average t</summary>
+            public double averageT;
+
+            /// <summary>The sunshine hours</summary>
+            public double sunshineHours;
+
+            /// <summary>The fraction clear sky</summary>
+            public double fractionClearSky;
+
+            /// <summary>The day length</summary>
+            public double dayLength;
+
+            /// <summary>The day length light</summary>
+            public double dayLengthLight;
+
+            /// <summary>The delta z</summary>
+            public double[] DeltaZ = new double[-1 + 1];
+
+            /// <summary>The layer ktot</summary>
+            public double[] layerKtot = new double[-1 + 1];
+
+            /// <summary>The layer la isum</summary>
+            public double[] layerLAIsum = new double[-1 + 1];
+
+            /// <summary>The number layers</summary>
+            public int numLayers;
+
+            /// <summary>The soil_heat</summary>
+            public double soil_heat = 0;
+
+            /// <summary>The dryleaffraction</summary>
+            public double dryleaffraction = 0;
+
+            /// <summary>The emissivity</summary>
+            public double Emissivity = 0.96;
+
+            /// <summary>Gets or sets the component data.</summary>
+            public List<CanopyType> Canopies = new List<CanopyType>();
+
+            /// <summary>Break the combined Canopy into layers</summary>
+            public void DefineLayers()
+            {
+                double[] nodes = new double[2 * Canopies.Count];
+                int numNodes = 1;
+                for (int compNo = 0; compNo <= Canopies.Count - 1; compNo++)
+                {
+                    double height = Canopies[compNo].HeightMetres;
+                    double canopyBase = height - Canopies[compNo].DepthMetres;
+                    if (Array.IndexOf(nodes, height) == -1)
+                    {
+                        nodes[numNodes] = height;
+                        numNodes = numNodes + 1;
+                    }
+                    if (Array.IndexOf(nodes, canopyBase) == -1)
+                    {
+                        nodes[numNodes] = canopyBase;
+                        numNodes = numNodes + 1;
+                    }
+                }
+                Array.Resize<double>(ref nodes, numNodes);
+                Array.Sort(nodes);
+                numLayers = numNodes - 1;
+                if (DeltaZ.Length != numLayers)
+                {
+                    // Number of layers has changed; adjust array lengths
+                    Array.Resize<double>(ref DeltaZ, numLayers);
+                    Array.Resize<double>(ref layerKtot, numLayers);
+                    Array.Resize<double>(ref layerLAIsum, numLayers);
+
+                    for (int j = 0; j <= Canopies.Count - 1; j++)
+                    {
+                        Array.Resize<double>(ref Canopies[j].Ftot, numLayers);
+                        Array.Resize<double>(ref Canopies[j].Fgreen, numLayers);
+                        Array.Resize<double>(ref Canopies[j].Rs, numLayers);
+                        Array.Resize<double>(ref Canopies[j].Rl, numLayers);
+                        Array.Resize<double>(ref Canopies[j].Rsoil, numLayers);
+                        Array.Resize<double>(ref Canopies[j].Gc, numLayers);
+                        Array.Resize<double>(ref Canopies[j].Ga, numLayers);
+                        Array.Resize<double>(ref Canopies[j].PET, numLayers);
+                        Array.Resize<double>(ref Canopies[j].PETr, numLayers);
+                        Array.Resize<double>(ref Canopies[j].PETa, numLayers);
+                        Array.Resize<double>(ref Canopies[j].Omega, numLayers);
+                        Array.Resize<double>(ref Canopies[j].interception, numLayers);
+                    }
+                }
+                for (int i = 0; i <= numNodes - 2; i++)
+                {
+                    DeltaZ[i] = nodes[i + 1] - nodes[i];
+                }
+            }
+
+            /// <summary>Break the components into layers</summary>
+            public void DivideComponents()
+            {
+                double[] Ld = new double[Canopies.Count];
+                for (int j = 0; j <= Canopies.Count - 1; j++)
+                {
+                    CanopyType componentData = Canopies[j];
+
+                    componentData.layerLAI = new double[numLayers];
+                    componentData.layerLAItot = new double[numLayers];
+                    Ld[j] = MathUtilities.Divide(componentData.Canopy.LAITotal, componentData.DepthMetres, 0.0);
+                }
+                double top = 0.0;
+                double bottom = 0.0;
+
+                for (int i = 0; i <= numLayers - 1; i++)
+                {
+                    bottom = top;
+                    top = top + DeltaZ[i];
+                    layerLAIsum[i] = 0.0;
+
+                    // Calculate LAI for layer i and component j
+                    // ===========================================
+                    for (int j = 0; j <= Canopies.Count - 1; j++)
+                    {
+                        CanopyType componentData = Canopies[j];
+
+                        if ((componentData.HeightMetres > bottom) && (componentData.HeightMetres - componentData.DepthMetres < top))
+                        {
+                            componentData.layerLAItot[i] = Ld[j] * DeltaZ[i];
+                            componentData.layerLAI[i] = componentData.layerLAItot[i] * MathUtilities.Divide(componentData.Canopy.LAI, componentData.Canopy.LAITotal, 0.0);
+                            layerLAIsum[i] += componentData.layerLAItot[i];
+                        }
+                    }
+
+                    // Calculate fractional contribution for layer i and component j
+                    // ====================================================================
+                    for (int j = 0; j <= Canopies.Count - 1; j++)
+                    {
+                        CanopyType componentData = Canopies[j];
+
+                        componentData.Ftot[i] = MathUtilities.Divide(componentData.layerLAItot[i], layerLAIsum[i], 0.0);
+                        // Note: Sum of Fgreen will be < 1 as it is green over total
+                        componentData.Fgreen[i] = MathUtilities.Divide(componentData.layerLAI[i], layerLAIsum[i], 0.0);
+                    }
+                }
+            }
+
+            /// <summary>
+            /// Reset class state
+            /// </summary>
+            public void Reset()
+            {
+                soil_heat = 0.0;
+                dryleaffraction = 0.0;
+                _albedo = 0.0;// albedo;
+                netLongWave = 0;
+                sumRs = 0;
+                averageT = 0;
+                sunshineHours = 0;
+                fractionClearSky = 0;
+                dayLength = 0;
+                dayLengthLight = 0;
+                numLayers = 0;
+                DeltaZ = new double[-1 + 1];
+                layerKtot = new double[-1 + 1];
+                layerLAIsum = new double[-1 + 1];
+                Canopies.Clear();
+
+            }
+
+        }
+    }
+}
